Minor fixes for python3

+++ utils.py
@@ -31,7 +31,7 @@ def shorten(string, dx, text_size):
     # http://stackoverflow.com/questions/1123463/clipping-text-in-python-tkinter
 
     if (len(string)-5) * text_size > dx:
-        newLen = dx / text_size + 5
+        newLen = int(dx / text_size + 5)
         string = string[0:newLen - 1]
 
     return string
